services/ocr-engine/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pytesseract>=0.3.8
pydantic>=2.0.0
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiofiles==23.1.0
requests==2.31.0
sse-starlette>=1.6.0
prometheus-client>=0.17.0
Pillow>=10.0.0  # For image processing
